tests: shared integration harness and single launcher layout
Restructure tests/integration as the home of the integration suites,
with run.sh as the single entrypoint at the top:

  tests/integration/{README.md, run.sh, runner.yaml,
                     harness/, snapshot-agent/}

The new harness package holds the generic pieces extracted from the
snapshot-agent harness — in-cluster client setup, node selection
(TEST_NODE / free-GPU pick), pod lifecycle, exec (with stdin) / HTTP /
VRAM helpers, and attach-by-label for chart-deployed pods. The
snapshot-agent suite keeps its specifics (agent deployment, engine
specs, agentctl call path) and composes the shared Cluster.

Pure refactor: no behavior change.
