+++ .github/workflows/label-integration-test.yaml
@@ -48,7 +48,7 @@ jobs:
           labels=$(gh api "repos/$REPO/issues/$PR/labels" --jq '.[].name')
           if printf '%s\n' "$labels" | grep -Fxq "needs-integration-test"; then
             echo "::error::This PR modifies snapshot-agent and requires manual integration testing."
-            echo "::error::See tests/integration/snapshot-agent/README.md for how to run the integration tests."
+            echo "::error::See tests/integration/README.md for how to run the integration tests."
             echo "::error::Post the test results as a comment on this PR, or if integration tests are not applicable, comment a justification. Then remove the 'needs-integration-test' label."
             exit 1
           fi
