+++ tests/integration/snapshot-agent/harness.go
@@ -8,8 +8,14 @@
 // the production path for workloads — never a Go gRPC client. The harness
 // talks to the engines over HTTP directly for inference checks.
 //
-// Test cases live in standalone_test.go and k8s_test.go. To add a test, add
-// a t.Run(...) inside the engine group that provides the pods it needs.
+// Generic cluster helpers (client setup, node selection, pod lifecycle,
+// exec) live in tests/integration/harness; this package adds the
+// snapshot-agent specifics: agent deployment, engine specs, and the
+// agentctl.py call path.
+//
+// Test cases live in standalone_test.go, k8s_test.go and
+// orchestrator_test.go. To add a test, add a t.Run(...) inside the engine
+// group that provides the pods it needs.
 package integration
 
 import (
@@ -28,24 +34,20 @@ import (
 	"testing"
 	"time"
 
-	corev1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/remotecommand"
+
+	"github.com/llm-d-incubation/llm-d-rl-time-slicing/tests/integration/harness"
 )
 
 const (
-	namespace     = "default"
-	agentPodName  = "snapshot-agent-test"
-	agentPort     = 9001
+	namespace    = "default"
+	agentPodName = "snapshot-agent-test"
+	agentPort    = 9001
 	// chartNamespace is where the official Helm charts install components:
 	// their templates pin the namespace rather than using the release's.
 	chartNamespace = "timeslice-system"
-	podTimeout    = 5 * time.Minute
-	healthTimeout = 5 * time.Minute
+	podTimeout     = 5 * time.Minute
+	healthTimeout  = 5 * time.Minute
 	// opTimeout bounds one agentctl.py invocation (RPC + operation polling);
 	// it also guards against the client's unbounded wait_for_operation.
 	opTimeout = 120 * time.Second
@@ -55,11 +57,10 @@ const (
 
 // Harness manages the test stack for one deployment mode.
 type Harness struct {
-	Client *kubernetes.Clientset
-	Config *rest.Config
-	Node   string
-	Model  string
-	Mode   string // "standalone" or "k8s"
+	*harness.Cluster
+	Node  string
+	Model string
+	Mode  string // "standalone" or "k8s"
 
 	AgentIP string
 	// AgentPort is the agent's gRPC port: agentPort for harness-deployed
@@ -74,28 +75,20 @@ type Harness struct {
 func NewHarness(t *testing.T) *Harness {
 	t.Helper()
 
-	cfg, err := rest.InClusterConfig()
-	if err != nil {
-		t.Fatalf("in-cluster config: %v", err)
-	}
-	client, err := kubernetes.NewForConfig(cfg)
-	if err != nil {
-		t.Fatalf("k8s client: %v", err)
-	}
-
 	model := os.Getenv("MODEL")
 	if model == "" {
 		model = "Qwen/Qwen2.5-0.5B"
 	}
 
-	h := &Harness{Client: client, Config: cfg, Model: model, Mode: "standalone", AgentPort: agentPort}
+	h := &Harness{
+		Cluster:   harness.NewCluster(t, namespace),
+		Model:     model,
+		Mode:      "standalone",
+		AgentPort: agentPort,
+	}
 	// TEST_NODE pins the suite to a specific node (e.g. one known to be
 	// otherwise idle); by default the first node with a free GPU is used.
-	if node := os.Getenv("TEST_NODE"); node != "" {
-		h.Node = node
-	} else {
-		h.Node = h.pickGPUNode(t)
-	}
+	h.Node = h.PickNode(t)
 	t.Logf("using node %s, mode %s", h.Node, h.Mode)
 
 	h.deployAgent(t)
@@ -109,25 +102,19 @@ func NewHarness(t *testing.T) *Harness {
 func NewChartHarness(t *testing.T) *Harness {
 	t.Helper()
 
-	cfg, err := rest.InClusterConfig()
-	if err != nil {
-		t.Fatalf("in-cluster config: %v", err)
-	}
-	client, err := kubernetes.NewForConfig(cfg)
-	if err != nil {
-		t.Fatalf("k8s client: %v", err)
-	}
-
-	node := os.Getenv("TEST_NODE")
-	if node == "" {
-		t.Fatal("TEST_NODE env var is required for chart tests")
-	}
+	node := harness.RequiredNode(t)
 	model := os.Getenv("MODEL")
 	if model == "" {
 		model = "Qwen/Qwen2.5-0.5B"
 	}
 
-	h := &Harness{Client: client, Config: cfg, Model: model, Mode: "k8s", Node: node, AgentPort: agentPort}
+	h := &Harness{
+		Cluster:   harness.NewCluster(t, namespace),
+		Model:     model,
+		Mode:      "k8s",
+		Node:      node,
+		AgentPort: agentPort,
+	}
 	if p := os.Getenv("CHART_AGENT_PORT"); p != "" {
 		port, err := strconv.Atoi(p)
 		if err != nil {
@@ -135,102 +122,26 @@ func NewChartHarness(t *testing.T) *Harness {
 		}
 		h.AgentPort = port
 	}
-	h.AgentIP = h.waitChartAgentReady(t)
+	// The chart runs the agent on hostNetwork, so the pod IP is the node IP —
+	// also how the orchestrator reaches agents.
+	h.AgentIP = h.WaitPodReadyByLabel(t, chartNamespace,
+		"app.kubernetes.io/name=snapshot-agent,app.kubernetes.io/instance=sa-chart-test",
+		h.Node, podTimeout)
 	t.Logf("using node %s, chart-deployed agent at %s:%d", h.Node, h.AgentIP, h.AgentPort)
 	return h
 }
 
-// waitChartAgentReady waits for the chart's DaemonSet pod on the harness node
-// and returns its IP (the node IP — the chart runs the agent on hostNetwork,
-// which is also how the orchestrator reaches agents).
-func (h *Harness) waitChartAgentReady(t *testing.T) string {
-	t.Helper()
-	deadline := time.Now().Add(podTimeout)
-	for time.Now().Before(deadline) {
-		pods, err := h.Client.CoreV1().Pods(chartNamespace).List(context.Background(), metav1.ListOptions{
-			LabelSelector: "app.kubernetes.io/name=snapshot-agent,app.kubernetes.io/instance=sa-chart-test",
-			FieldSelector: "spec.nodeName=" + h.Node,
-		})
-		if err == nil {
-			for i := range pods.Items {
-				pod := &pods.Items[i]
-				for _, cond := range pod.Status.Conditions {
-					if cond.Type == corev1.PodReady && cond.Status == corev1.ConditionTrue {
-						return pod.Status.PodIP
-					}
-				}
-			}
-		}
-		time.Sleep(3 * time.Second)
-	}
-	t.Fatalf("timeout waiting for chart-deployed snapshot-agent on node %s in namespace %s", h.Node, chartNamespace)
-	return ""
-}
-
-// pickGPUNode returns the first node with at least one FREE GPU
-// (allocatable minus what running pods have requested). Engines run one at a
-// time, so a single free GPU is enough.
-func (h *Harness) pickGPUNode(t *testing.T) string {
-	t.Helper()
-	nodes, err := h.Client.CoreV1().Nodes().List(context.Background(), metav1.ListOptions{
-		LabelSelector: "cloud.google.com/gke-accelerator",
-	})
-	if err != nil {
-		t.Fatalf("listing GPU nodes: %v", err)
-	}
-	for i := range nodes.Items {
-		node := &nodes.Items[i]
-		alloc := node.Status.Allocatable[corev1.ResourceName("nvidia.com/gpu")]
-		if alloc.Value() < 1 {
-			continue
-		}
-		used, err := h.gpusRequestedOnNode(node.Name)
-		if err != nil {
-			t.Fatalf("counting GPU requests on %s: %v", node.Name, err)
-		}
-		if alloc.Value()-used >= 1 {
-			return node.Name
-		}
-		t.Logf("node %s has no free GPUs (%d/%d requested), skipping", node.Name, used, alloc.Value())
-	}
-	t.Fatal("no node with a free GPU found")
-	return ""
-}
-
-// gpusRequestedOnNode sums nvidia.com/gpu requests of non-terminated pods on
-// the node.
-func (h *Harness) gpusRequestedOnNode(node string) (int64, error) {
-	pods, err := h.Client.CoreV1().Pods("").List(context.Background(), metav1.ListOptions{
-		FieldSelector: "spec.nodeName=" + node,
-	})
-	if err != nil {
-		return 0, err
-	}
-	var used int64
-	for i := range pods.Items {
-		pod := &pods.Items[i]
-		if pod.Status.Phase == corev1.PodSucceeded || pod.Status.Phase == corev1.PodFailed {
-			continue
-		}
-		for j := range pod.Spec.Containers {
-			req := pod.Spec.Containers[j].Resources.Requests[corev1.ResourceName("nvidia.com/gpu")]
-			used += req.Value()
-		}
-	}
-	return used, nil
-}
-
 func (h *Harness) deployAgent(t *testing.T) {
 	t.Helper()
-	h.deletePodAndWait(t, agentPodName)
+	h.DeletePodAndWait(t, agentPodName)
 
 	pod := agentPod(h.Node)
 	if _, err := h.Client.CoreV1().Pods(namespace).Create(context.Background(), pod, metav1.CreateOptions{}); err != nil {
 		t.Fatalf("creating agent pod: %v", err)
 	}
-	t.Cleanup(func() { h.deletePodAndWait(t, agentPodName) })
+	t.Cleanup(func() { h.DeletePodAndWait(t, agentPodName) })
 
-	h.AgentIP = h.waitPodReady(t, agentPodName)
+	h.AgentIP = h.WaitPodReady(t, agentPodName, podTimeout)
 	h.installAgentBinaries(t)
 	h.waitAgentUp(t)
 	t.Logf("agent (make-standalone artifacts) ready at %s:%d", h.AgentIP, h.AgentPort)
@@ -265,10 +176,10 @@ func (h *Harness) installAgentBinaries(t *testing.T) {
 		t.Fatalf("closing tar stream: %v", err)
 	}
 
-	if _, err := h.execPodStdin(agentPodName, "snapshot-agent", &buf, "tar", "-xf", "-", "-C", "/opt/rlts"); err != nil {
+	if _, err := h.ExecPodStdin(agentPodName, "snapshot-agent", &buf, opTimeout, "tar", "-xf", "-", "-C", "/opt/rlts"); err != nil {
 		t.Fatalf("copying standalone artifacts into agent pod: %v", err)
 	}
-	if _, err := h.execPod(agentPodName, "snapshot-agent", "touch", "/opt/rlts/.ready"); err != nil {
+	if _, err := h.ExecPod(agentPodName, "snapshot-agent", opTimeout, "touch", "/opt/rlts/.ready"); err != nil {
 		t.Fatalf("releasing agent pod: %v", err)
 	}
 }
@@ -295,17 +206,17 @@ func (h *Harness) waitAgentUp(t *testing.T) {
 func (h *Harness) WithEngine(t *testing.T, spec EngineSpec, fn func(t *testing.T, e *Engine)) {
 	t.Helper()
 	podName := spec.Name + "-test"
-	h.deletePodAndWait(t, podName)
+	h.DeletePodAndWait(t, podName)
 
 	pod := spec.BuildPod(h)
 	if _, err := h.Client.CoreV1().Pods(namespace).Create(context.Background(), pod, metav1.CreateOptions{}); err != nil {
 		t.Fatalf("creating %s pod: %v", spec.Name, err)
 	}
-	defer h.deletePodAndWait(t, podName)
+	defer h.DeletePodAndWait(t, podName)
 
-	ip := h.waitPodReady(t, podName)
+	ip := h.WaitPodReady(t, podName, podTimeout)
 	t.Logf("%s pod ready at %s, waiting for model load...", spec.Name, ip)
-	h.waitHTTP(t, fmt.Sprintf("http://%s:%d/health", ip, spec.Port))
+	h.WaitHTTP(t, fmt.Sprintf("http://%s:%d/health", ip, spec.Port), healthTimeout)
 
 	e := &Engine{Spec: spec, IP: ip, PodName: podName}
 	if h.Mode == "standalone" {
@@ -334,111 +245,13 @@ func (e *Engine) Endpoint() string {
 	return fmt.Sprintf("http://%s:%d", e.IP, e.Spec.Port)
 }
 
-// --- Pod helpers ---
-
-func (h *Harness) waitPodReady(t *testing.T, name string) string {
-	t.Helper()
-	deadline := time.Now().Add(podTimeout)
-	for time.Now().Before(deadline) {
-		pod, err := h.Client.CoreV1().Pods(namespace).Get(context.Background(), name, metav1.GetOptions{})
-		if err == nil {
-			for _, cond := range pod.Status.Conditions {
-				if cond.Type == corev1.PodReady && cond.Status == corev1.ConditionTrue {
-					return pod.Status.PodIP
-				}
-			}
-		}
-		time.Sleep(3 * time.Second)
-	}
-	t.Fatalf("timeout waiting for pod %s to become ready", name)
-	return ""
-}
-
-func (h *Harness) deletePodAndWait(t *testing.T, name string) {
-	t.Helper()
-	zero := int64(0)
-	err := h.Client.CoreV1().Pods(namespace).Delete(context.Background(), name, metav1.DeleteOptions{
-		GracePeriodSeconds: &zero,
-	})
-	if apierrors.IsNotFound(err) {
-		return
-	}
-	if err != nil {
-		t.Fatalf("deleting pod %s: %v", name, err)
-	}
-	deadline := time.Now().Add(2 * time.Minute)
-	for time.Now().Before(deadline) {
-		_, err := h.Client.CoreV1().Pods(namespace).Get(context.Background(), name, metav1.GetOptions{})
-		if apierrors.IsNotFound(err) {
-			return
-		}
-		time.Sleep(2 * time.Second)
-	}
-	t.Fatalf("timeout waiting for pod %s to be deleted", name)
-}
-
-// waitHTTP polls url until it returns 200. Pod readiness only means the
-// container started; engines need additional time to load the model.
-func (h *Harness) waitHTTP(t *testing.T, url string) {
-	t.Helper()
-	client := &http.Client{Timeout: 5 * time.Second}
-	deadline := time.Now().Add(healthTimeout)
-	for time.Now().Before(deadline) {
-		resp, err := client.Get(url)
-		if err == nil {
-			resp.Body.Close()
-			if resp.StatusCode == http.StatusOK {
-				return
-			}
-		}
-		time.Sleep(5 * time.Second)
-	}
-	t.Fatalf("timeout waiting for %s", url)
-}
-
-// execPod runs a command in a pod container and returns stdout.
-func (h *Harness) execPod(pod, container string, command ...string) (string, error) {
-	return h.execPodStdin(pod, container, nil, command...)
-}
-
-// execPodStdin runs a command in a pod container, streaming stdin into it if
-// non-nil, and returns stdout.
-func (h *Harness) execPodStdin(pod, container string, stdin io.Reader, command ...string) (string, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), opTimeout)
-	defer cancel()
-	req := h.Client.CoreV1().RESTClient().Post().
-		Resource("pods").Name(pod).Namespace(namespace).SubResource("exec").
-		VersionedParams(&corev1.PodExecOptions{
-			Container: container,
-			Command:   command,
-			Stdin:     stdin != nil,
-			Stdout:    true,
-			Stderr:    true,
-		}, scheme.ParameterCodec)
-
-	exec, err := remotecommand.NewSPDYExecutor(h.Config, "POST", req.URL())
-	if err != nil {
-		return "", fmt.Errorf("creating executor: %w", err)
-	}
-	var stdout, stderr bytes.Buffer
-	err = exec.StreamWithContext(ctx, remotecommand.StreamOptions{
-		Stdin:  stdin,
-		Stdout: &stdout,
-		Stderr: &stderr,
-	})
-	if err != nil {
-		return "", fmt.Errorf("exec failed: %w (stderr: %s)", err, stderr.String())
-	}
-	return stdout.String(), nil
-}
-
 // findPID locates the engine process on the node via the agent pod (hostPID).
 func (h *Harness) findPID(t *testing.T, pattern string) int32 {
 	t.Helper()
 	script := fmt.Sprintf(
 		`for p in /proc/[0-9]*/cmdline; do grep -ql '%s' $p 2>/dev/null && echo $p | cut -d/ -f3 && break; done`,
 		pattern)
-	out, err := h.execPod(agentPodName, "snapshot-agent", "sh", "-c", script)
+	out, err := h.ExecPod(agentPodName, "snapshot-agent", opTimeout, "sh", "-c", script)
 	if err != nil {
 		t.Fatalf("finding PID for %q: %v", pattern, err)
 	}
@@ -452,16 +265,7 @@ func (h *Harness) findPID(t *testing.T, pattern string) int32 {
 // VRAMMiB returns the GPU memory used (MiB) as seen from the engine pod.
 func (h *Harness) VRAMMiB(t *testing.T, e *Engine) int {
 	t.Helper()
-	out, err := h.execPod(e.PodName, e.Spec.Name,
-		"nvidia-smi", "--query-gpu=memory.used", "--format=csv,noheader,nounits")
-	if err != nil {
-		t.Fatalf("querying VRAM on %s: %v", e.PodName, err)
-	}
-	mib, err := strconv.Atoi(strings.TrimSpace(strings.Split(out, "\n")[0]))
-	if err != nil {
-		t.Fatalf("parsing VRAM from %q: %v", out, err)
-	}
-	return mib
+	return h.PodVRAMMiB(t, e.PodName, e.Spec.Name, opTimeout)
 }
 
 // --- Agent call helpers ---
@@ -577,4 +381,3 @@ func RequireFreedAndCorrect(t *testing.T, vramWhileAsleep int, before, after str
 		t.Errorf("inference changed after restore: before=%q after=%q", before, after)
 	}
 }
-
